+++ nni/experiment/__init__.py
@@ -2,8 +2,5 @@
 # Licensed under the MIT license.
 
 from .config import *
-
-# FIXME: temporarily disabled to run test case
-
-#from .experiment import Experiment
-#from .data import *
+from .experiment import Experiment, RunMode
+from .data import *
